Users can only delete their credentials

+++ lib/lightning_web/live/credential_live/index.ex
@@ -5,31 +5,16 @@ defmodule LightningWeb.CredentialLive.Index do
   use LightningWeb, :live_view
 
   alias Lightning.Credentials
-  alias Lightning.Policies.{Users, Permissions}
 
   @impl true
   def mount(_params, _session, socket) do
-    can_access_own_credentials =
-      Users
-      |> Permissions.can(
-        :access_own_credentials,
-        socket.assigns.current_user,
-        socket.assigns.current_user
-      )
-
-    if can_access_own_credentials do
-      {:ok,
-       assign(
-         socket,
-         :credentials,
-         list_credentials(socket.assigns.current_user.id)
-       )
-       |> assign(:active_menu_item, :credentials)}
-    else
-      {:ok,
-       put_flash(socket, :error, "You can't access that page")
-       |> push_redirect(to: "/")}
-    end
+    {:ok,
+     assign(
+       socket,
+       :credentials,
+       list_credentials(socket.assigns.current_user.id)
+     )
+     |> assign(:active_menu_item, :credentials)}
   end
 
   @impl true
@@ -47,19 +32,33 @@ defmodule LightningWeb.CredentialLive.Index do
   def handle_event("delete", %{"id" => id}, socket) do
     credential = Credentials.get_credential!(id)
 
-    Credentials.delete_credential(credential)
-    |> case do
-      {:ok, _} ->
-        {:noreply,
-         socket
-         |> assign(
-           :credentials,
-           list_credentials(socket.assigns.current_user.id)
-         )
-         |> put_flash(:info, "Credential deleted successfully")}
+    can_delete_credential =
+      Lightning.Policies.Users
+      |> Lightning.Policies.Permissions.can(
+        :delete_credential,
+        socket.assigns.current_user,
+        credential
+      )
 
-      {:error, _changeset} ->
-        {:noreply, socket |> put_flash(:error, "Can't delete credential")}
+    if can_delete_credential do
+      Credentials.delete_credential(credential)
+      |> case do
+        {:ok, _} ->
+          {:noreply,
+           socket
+           |> assign(
+             :credentials,
+             list_credentials(socket.assigns.current_user.id)
+           )
+           |> put_flash(:info, "Credential deleted successfully")}
+
+        {:error, _changeset} ->
+          {:noreply, socket |> put_flash(:error, "Can't delete credential")}
+      end
+    else
+      {:noreply,
+       put_flash(socket, :error, "You can't perform this action")
+       |> push_patch(to: ~p"/credentials")}
     end
   end
